Moved TestObserver out for access across tests.
